Add cubic interactive
