+++ public/content/lessons/2021/quintic/index.mdx
@@ -10,8 +10,18 @@ credits:
 
 There is no text version of this lesson yet. But in the meantime, enjoy the following interactive roots finder.
 
+# Quintic
+
 <Interactive
-  filename="quintic-interactive/index"
+  filename="quintic-interactive/quintic"
   aspectRatio={640 / 420}
+  allowFullscreen={true}
+/>
+
+# Cubic
+
+<Interactive
+  filename="quintic-interactive/cubic"
+  aspectRatio={640 / 540}
   allowFullscreen={true}
 />
